feat(refactor): Move common funtion to Common Utils
