+++ src/utils/common.utils.ts
@@ -8,6 +8,9 @@
 // SPDX-License-Identifier: EPL-2.0
 // License-Filename: LICENSE
 
+import InputKeyValue from '@/object-types/InputKeyValue'
+import Moderators from '@/object-types/Moderators'
+
 const isNullOrUndefined = (obj: unknown) => {
     if (obj === null || obj === undefined) {
         return true
@@ -46,12 +49,121 @@ const getIdFromUrl = (url: string): string => {
     return url.split('/').at(-1)
 }
 
+const handlerModerators = (emails: any[]) => {
+    const fullNames: string[] = []
+    const moderatorsEmail: string[] = []
+    if (emails.length == 0) {
+        return
+    }
+    emails.forEach((item) => {
+        fullNames.push(item.fullName)
+        moderatorsEmail.push(item.email)
+    })
+    const moderatorsName: string = fullNames.join(' , ')
+    const moderatorsResponse: Moderators = {
+        fullName: moderatorsName,
+        emails: moderatorsEmail,
+    }
+    return moderatorsResponse
+}
+
+const handlerContributor = (emails: any[]) => {
+    const fullNames: string[] = []
+    const contributorsEmail: string[] = []
+    if (emails.length == 0) {
+        return
+    }
+    emails.forEach((item) => {
+        fullNames.push(item.fullName)
+        contributorsEmail.push(item.email)
+    })
+    const contributorsName: string = fullNames.join(' , ')
+    const contributorsResponse: Moderators = {
+        fullName: contributorsName,
+        emails: contributorsEmail,
+    }
+    return contributorsResponse
+}
+
+const getEmailsModerators = (emails: any[]) => {
+    const moderatorsEmail: string[] = []
+    if (typeof emails === 'undefined') {
+        return
+    }
+    emails.forEach((item) => {
+        moderatorsEmail.push(item.email)
+    })
+
+    return moderatorsEmail
+}
+
+const convertObjectToMap = (data: string) => {
+    const map = new Map(Object.entries(data))
+    const inputs: InputKeyValue[] = []
+    map.forEach((value, key) => {
+        const input: InputKeyValue = {
+            key: key,
+            value: value,
+        }
+        inputs.push(input)
+    })
+    return inputs
+}
+
+const convertObjectToMapRoles = (data: string) => {
+    if (data === undefined) {
+        return null
+    }
+    const inputRoles: InputKeyValue[] = []
+    const mapRoles = new Map(Object.entries(data))
+    mapRoles.forEach((value, key) => {
+        for (let index = 0; index < value.length; index++) {
+            const input: InputKeyValue = {
+                key: key,
+                value: value.at(index),
+            }
+            inputRoles.push(input)
+        }
+    })
+    return inputRoles
+}
+
+const convertRoles = (datas: InputKeyValue[]) => {
+    if (datas === null) {
+        return null
+    }
+    const contributors: string[] = []
+    const commiters: string[] = []
+    const expecters: string[] = []
+    datas.forEach((data) => {
+        if (data.key === 'Contributor') {
+            contributors.push(data.value)
+        } else if (data.key === 'Committer') {
+            commiters.push(data.value)
+        } else if (data.key === 'Expert') {
+            expecters.push(data.value)
+        }
+    })
+    const roles = {
+        Contributor: contributors,
+        Committer: commiters,
+        Expert: expecters,
+    }
+    return roles
+}
+
 const CommonUtils = {
     isNullOrUndefined,
     isNullEmptyOrUndefinedString,
     createUrlWithParams,
     isNullEmptyOrUndefinedArray,
     getIdFromUrl,
+    handlerModerators,
+    handlerContributor,
+    getEmailsModerators,
+    convertObjectToMap,
+    convertObjectToMapRoles,
+    convertRoles,
 }
 
 export default CommonUtils
